Move checkout related fixtures from global fixture file
Move checkout related fixtures from other files

Destroy benchamrk conftest

Fix lint; move ommmited order fixture
