+++ conftest.py
@@ -25,10 +25,8 @@ prepare_test_db_connections()
 pytest_plugins = [
     "saleor.tests.fixtures",
     "saleor.app.tests.fixtures",
-    "saleor.discount.tests.test_utils.fixtures",
     "saleor.plugins.tests.fixtures",
     "saleor.graphql.tests.fixtures",
-    "saleor.graphql.checkout.tests.benchmark.fixtures",
     "saleor.graphql.account.tests.benchmark.fixtures",
     "saleor.webhook.tests.fixtures",
     "saleor.tax.tests.fixtures",
@@ -44,6 +42,7 @@ pytest_plugins = [
     "saleor.permission.tests.fixtures",
     "saleor.giftcard.tests.fixtures",
     "saleor.discount.tests.fixtures",
+    "saleor.checkout.tests.fixtures",
 ]
