+++ saleor/checkout/tests/fixtures/__init__.py
@@ -1,0 +1,9 @@
+from .benchmark import *  # noqa: F403
+from .checkout import *  # noqa: F403
+from .checkout_info import *  # noqa: F403
+from .checkout_line import *  # noqa: F403
+from .checkout_line_info import *  # noqa: F403
+from .checkout_with_discount import *  # noqa: F403
+from .checkout_with_payment import *  # noqa: F403
+from .checkout_with_prices import *  # noqa: F403
+from .checkout_with_transaction import *  # noqa: F403
