+++ saleor/tests/fixtures.py
@@ -14,10 +14,8 @@ from django.core.files.uploadedfile import SimpleUploadedFile
 from django.db import connection
 from django.template.defaultfilters import truncatechars
 from django.test.utils import CaptureQueriesContext as BaseCaptureQueriesContext
-from django.utils import timezone
 from freezegun import freeze_time
 from PIL import Image
-from prices import Money
 
 from ..account.models import Address, Group, StaffNotificationRecipient, User
 from ..attribute import AttributeEntityType, AttributeInputType, AttributeType
@@ -28,13 +26,6 @@ from ..attribute.models import (
     AttributeValueTranslation,
 )
 from ..attribute.utils import associate_attribute_values_to_instance
-from ..checkout import base_calculations
-from ..checkout.fetch import fetch_checkout_info, fetch_checkout_lines
-from ..checkout.models import Checkout, CheckoutLine, CheckoutMetadata
-from ..checkout.utils import (
-    add_variant_to_checkout,
-    add_voucher_to_checkout,
-)
 from ..core import JobStatus
 from ..core.models import EventDelivery, EventDeliveryAttempt, EventPayload
 from ..core.payments import PaymentInterface
@@ -43,16 +34,9 @@ from ..core.utils.editorjs import clean_editor_js
 from ..csv.events import ExportEvents
 from ..csv.models import ExportEvent, ExportFile
 from ..discount import (
-    DiscountType,
-    DiscountValueType,
     PromotionEvents,
-    RewardType,
-    RewardValueType,
 )
 from ..discount.models import (
-    CheckoutDiscount,
-    CheckoutLineDiscount,
-    Promotion,
     PromotionEvent,
     PromotionRuleTranslation,
     PromotionTranslation,
@@ -64,19 +48,13 @@ from ..payment.models import Payment, TransactionEvent, TransactionItem
 from ..payment.transaction_item_calculations import recalculate_transaction_amounts
 from ..payment.utils import create_manual_adjustment_events
 from ..permission.enums import get_permissions
-from ..plugins.manager import get_plugins_manager
 from ..product.models import (
     CategoryTranslation,
     CollectionTranslation,
     ProductTranslation,
-    ProductVariantChannelListing,
     ProductVariantTranslation,
 )
 from ..tax import TaxCalculationStrategy
-from ..warehouse.models import (
-    PreorderReservation,
-    Reservation,
-)
 from ..webhook.event_types import WebhookEventAsyncType
 from ..webhook.transport.utils import to_payment_app_id
 from .utils import dummy_editorjs
@@ -162,707 +140,6 @@ def assert_num_queries(capture_queries):
 @pytest.fixture
 def assert_max_num_queries(capture_queries):
     return partial(capture_queries, exact=False)
-
-
-@pytest.fixture
-def checkout(db, channel_USD, settings):
-    checkout = Checkout.objects.create(
-        currency=channel_USD.currency_code,
-        channel=channel_USD,
-        price_expiration=timezone.now() + settings.CHECKOUT_PRICES_TTL,
-        email="user@email.com",
-    )
-    checkout.set_country("US", commit=True)
-    CheckoutMetadata.objects.create(checkout=checkout)
-    return checkout
-
-
-@pytest.fixture
-def checkout_JPY(channel_JPY):
-    checkout = Checkout.objects.create(
-        currency=channel_JPY.currency_code, channel=channel_JPY
-    )
-    checkout.set_country("JP", commit=True)
-    return checkout
-
-
-@pytest.fixture
-def checkout_with_item(checkout, product):
-    variant = product.variants.first()
-    checkout_info = fetch_checkout_info(
-        checkout, [], get_plugins_manager(allow_replica=False)
-    )
-    add_variant_to_checkout(checkout_info, variant, 3)
-    checkout.save()
-    return checkout
-
-
-@pytest.fixture
-def checkout_with_item_on_sale(checkout_with_item, promotion_converted_from_sale):
-    line = checkout_with_item.lines.first()
-    channel = checkout_with_item.channel
-    discount_amount = Decimal("5.0")
-    variant = line.variant
-    variant_id = graphene.Node.to_global_id("ProductVariant", variant.id)
-    predicate = {"variantPredicate": {"ids": [variant_id]}}
-    rule = promotion_converted_from_sale.rules.first()
-    rule.catalogue_predicate = predicate
-    rule.reward_value = discount_amount
-    rule.save(update_fields=["catalogue_predicate", "reward_value"])
-    rule.channels.add(channel)
-    channel_listing = variant.channel_listings.get(channel=channel)
-    channel_listing.discounted_price_amount = (
-        channel_listing.price_amount - discount_amount
-    )
-    channel_listing.save(update_fields=["discounted_price_amount"])
-
-    CheckoutLineDiscount.objects.create(
-        line=line,
-        promotion_rule=rule,
-        type=DiscountType.SALE,
-        value_type=rule.reward_value_type,
-        value=discount_amount,
-        amount_value=discount_amount * line.quantity,
-        currency=channel.currency_code,
-    )
-
-    return checkout_with_item
-
-
-@pytest.fixture
-def checkout_with_item_on_promotion(checkout_with_item):
-    line = checkout_with_item.lines.first()
-    channel = checkout_with_item.channel
-    promotion = Promotion.objects.create(name="Checkout promotion")
-
-    variant = line.variant
-
-    reward_value = Decimal("5")
-    rule = promotion.rules.create(
-        catalogue_predicate={
-            "productPredicate": {
-                "ids": [graphene.Node.to_global_id("Product", variant.product.id)]
-            }
-        },
-        reward_value_type=RewardValueType.FIXED,
-        reward_value=reward_value,
-    )
-    rule.channels.add(channel)
-
-    variant_channel_listing = variant.channel_listings.get(channel=channel)
-
-    variant_channel_listing.discounted_price_amount = (
-        variant_channel_listing.price_amount - reward_value
-    )
-    variant_channel_listing.save(update_fields=["discounted_price_amount"])
-
-    variant_channel_listing.variantlistingpromotionrule.create(
-        promotion_rule=rule,
-        discount_amount=reward_value,
-        currency=channel.currency_code,
-    )
-    CheckoutLineDiscount.objects.create(
-        line=line,
-        type=DiscountType.PROMOTION,
-        value_type=DiscountValueType.FIXED,
-        value=reward_value,
-        amount_value=reward_value * line.quantity,
-        currency=channel.currency_code,
-        promotion_rule=rule,
-    )
-
-    return checkout_with_item
-
-
-@pytest.fixture
-def checkout_with_item_and_order_discount(
-    checkout_with_item, catalogue_promotion_without_rules
-):
-    channel = checkout_with_item.channel
-
-    reward_value = Decimal("5")
-
-    rule = catalogue_promotion_without_rules.rules.create(
-        order_predicate={
-            "discountedObjectPredicate": {"baseSubtotalPrice": {"range": {"gte": 20}}}
-        },
-        reward_value_type=RewardValueType.FIXED,
-        reward_value=reward_value,
-        reward_type=RewardType.SUBTOTAL_DISCOUNT,
-    )
-    rule.channels.add(channel)
-
-    CheckoutDiscount.objects.create(
-        checkout=checkout_with_item,
-        promotion_rule=rule,
-        type=DiscountType.ORDER_PROMOTION,
-        value_type=rule.reward_value_type,
-        value=rule.reward_value,
-        amount_value=rule.reward_value,
-        currency=channel.currency_code,
-    )
-    checkout_with_item.discount_amount = reward_value
-    checkout_with_item.save(update_fields=["discount_amount"])
-
-    return checkout_with_item
-
-
-@pytest.fixture
-def checkout_with_item_and_gift_promotion(checkout_with_item, gift_promotion_rule):
-    channel = checkout_with_item.channel
-    variants = gift_promotion_rule.gifts.all()
-    variant_listings = ProductVariantChannelListing.objects.filter(variant__in=variants)
-    top_price, variant_id = max(
-        variant_listings.values_list("discounted_price_amount", "variant")
-    )
-
-    line = CheckoutLine.objects.create(
-        checkout=checkout_with_item,
-        quantity=1,
-        variant_id=variant_id,
-        is_gift=True,
-        currency="USD",
-    )
-
-    CheckoutLineDiscount.objects.create(
-        line=line,
-        promotion_rule=gift_promotion_rule,
-        type=DiscountType.ORDER_PROMOTION,
-        value_type=RewardValueType.FIXED,
-        value=top_price,
-        amount_value=top_price,
-        currency=channel.currency_code,
-    )
-
-    return checkout_with_item
-
-
-@pytest.fixture
-def checkout_with_item_and_transaction_item(checkout_with_item):
-    TransactionItem.objects.create(
-        name="Credit card",
-        psp_reference="PSP ref",
-        available_actions=["refund"],
-        currency="USD",
-        checkout_id=checkout_with_item.pk,
-        charged_value=Decimal("10"),
-    )
-    return checkout_with_item
-
-
-@pytest.fixture
-def checkout_with_item_and_tax_exemption(checkout_with_item):
-    checkout_with_item.tax_exemption = True
-    checkout_with_item.save(update_fields=["tax_exemption"])
-    return checkout_with_item
-
-
-@pytest.fixture
-def checkout_with_same_items_in_multiple_lines(checkout, product):
-    variant = product.variants.first()
-    checkout_info = fetch_checkout_info(
-        checkout, [], get_plugins_manager(allow_replica=False)
-    )
-    add_variant_to_checkout(checkout_info, variant, 1)
-    add_variant_to_checkout(checkout_info, variant, 1, force_new_line=True)
-    checkout.save()
-    return checkout
-
-
-@pytest.fixture
-def checkout_with_item_and_voucher_specific_products(
-    checkout_with_item, voucher_specific_product_type
-):
-    manager = get_plugins_manager(allow_replica=False)
-    lines, _ = fetch_checkout_lines(checkout_with_item)
-    checkout_info = fetch_checkout_info(checkout_with_item, lines, manager)
-    add_voucher_to_checkout(
-        manager,
-        checkout_info,
-        lines,
-        voucher_specific_product_type,
-        voucher_specific_product_type.codes.first(),
-    )
-    checkout_with_item.refresh_from_db()
-    return checkout_with_item
-
-
-@pytest.fixture
-def checkout_with_item_and_voucher_once_per_order(checkout_with_item, voucher):
-    voucher.apply_once_per_order = True
-    voucher.save()
-    manager = get_plugins_manager(allow_replica=False)
-    lines, _ = fetch_checkout_lines(checkout_with_item)
-    checkout_info = fetch_checkout_info(checkout_with_item, lines, manager)
-    add_voucher_to_checkout(
-        manager, checkout_info, lines, voucher, voucher.codes.first()
-    )
-    checkout_with_item.refresh_from_db()
-    return checkout_with_item
-
-
-@pytest.fixture
-def checkout_with_item_and_voucher(checkout_with_item, voucher):
-    manager = get_plugins_manager(allow_replica=False)
-    lines, _ = fetch_checkout_lines(checkout_with_item)
-    checkout_info = fetch_checkout_info(checkout_with_item, lines, manager)
-    add_voucher_to_checkout(
-        manager, checkout_info, lines, voucher, voucher.codes.first()
-    )
-    checkout_with_item.refresh_from_db()
-    return checkout_with_item
-
-
-@pytest.fixture
-def checkout_line(checkout_with_item):
-    return checkout_with_item.lines.first()
-
-
-@pytest.fixture
-def checkout_with_item_total_0(checkout, product_price_0):
-    variant = product_price_0.variants.get()
-    checkout_info = fetch_checkout_info(
-        checkout, [], get_plugins_manager(allow_replica=False)
-    )
-    add_variant_to_checkout(checkout_info, variant, 1)
-    checkout.save()
-    return checkout
-
-
-@pytest.fixture
-def checkout_JPY_with_item(checkout_JPY, product_in_channel_JPY):
-    variant = product_in_channel_JPY.variants.get()
-    checkout_info = fetch_checkout_info(
-        checkout_JPY, [], get_plugins_manager(allow_replica=False)
-    )
-    add_variant_to_checkout(checkout_info, variant, 3)
-    checkout_JPY.save()
-    return checkout_JPY
-
-
-@pytest.fixture
-def checkouts_list(channel_USD, channel_PLN):
-    checkouts_usd = Checkout.objects.bulk_create(
-        [
-            Checkout(currency=channel_USD.currency_code, channel=channel_USD),
-            Checkout(currency=channel_USD.currency_code, channel=channel_USD),
-            Checkout(currency=channel_USD.currency_code, channel=channel_USD),
-        ]
-    )
-    checkouts_pln = Checkout.objects.bulk_create(
-        [
-            Checkout(currency=channel_PLN.currency_code, channel=channel_PLN),
-            Checkout(currency=channel_PLN.currency_code, channel=channel_PLN),
-        ]
-    )
-    return [*checkouts_pln, *checkouts_usd]
-
-
-@pytest.fixture
-def checkouts_assigned_to_customer(channel_USD, channel_PLN, customer_user):
-    return Checkout.objects.bulk_create(
-        [
-            Checkout(
-                currency=channel_USD.currency_code,
-                channel=channel_USD,
-                user=customer_user,
-            ),
-            Checkout(
-                currency=channel_PLN.currency_code,
-                channel=channel_PLN,
-                user=customer_user,
-            ),
-        ]
-    )
-
-
-@pytest.fixture
-def checkout_ready_to_complete(checkout_with_item, address, shipping_method, gift_card):
-    checkout = checkout_with_item
-    checkout.shipping_address = address
-    checkout.shipping_method = shipping_method
-    checkout.billing_address = address
-    checkout.metadata_storage.store_value_in_metadata(items={"accepted": "true"})
-    checkout.metadata_storage.store_value_in_private_metadata(
-        items={"accepted": "false"}
-    )
-    checkout_with_item.gift_cards.add(gift_card)
-    checkout.save()
-    checkout.metadata_storage.save()
-    return checkout
-
-
-@pytest.fixture
-def checkout_with_digital_item(checkout, digital_content, address):
-    """Create a checkout with a digital line."""
-    variant = digital_content.product_variant
-    checkout_info = fetch_checkout_info(
-        checkout, [], get_plugins_manager(allow_replica=False)
-    )
-    add_variant_to_checkout(checkout_info, variant, 1)
-    checkout.discount_amount = Decimal(0)
-    checkout.billing_address = address
-    checkout.email = "customer@example.com"
-    checkout.save()
-    return checkout
-
-
-@pytest.fixture
-def checkout_with_shipping_required(checkout_with_item, product):
-    checkout = checkout_with_item
-    variant = product.variants.get()
-    checkout_info = fetch_checkout_info(
-        checkout, [], get_plugins_manager(allow_replica=False)
-    )
-    add_variant_to_checkout(checkout_info, variant, 3)
-    checkout.save()
-    return checkout
-
-
-@pytest.fixture
-def checkout_with_item_and_shipping_method(checkout_with_item, shipping_method):
-    checkout = checkout_with_item
-    checkout.shipping_method = shipping_method
-    checkout.save()
-    return checkout
-
-
-@pytest.fixture
-def checkout_with_item_and_voucher_and_shipping_method(
-    checkout_with_item_and_voucher, shipping_method
-):
-    checkout = checkout_with_item_and_voucher
-    checkout.shipping_method = shipping_method
-    checkout.save()
-    return checkout
-
-
-@pytest.fixture
-def checkout_without_shipping_required(checkout, product_without_shipping):
-    variant = product_without_shipping.variants.get()
-    checkout_info = fetch_checkout_info(
-        checkout, [], get_plugins_manager(allow_replica=False)
-    )
-    add_variant_to_checkout(checkout_info, variant, 1)
-    checkout.save()
-    return checkout
-
-
-@pytest.fixture
-def checkout_with_single_item(checkout, product):
-    variant = product.variants.get()
-    checkout_info = fetch_checkout_info(
-        checkout, [], get_plugins_manager(allow_replica=False)
-    )
-    add_variant_to_checkout(checkout_info, variant, 1)
-    checkout.save()
-    return checkout
-
-
-@pytest.fixture
-def checkout_with_variant_without_inventory_tracking(
-    checkout, variant_without_inventory_tracking, address, shipping_method
-):
-    variant = variant_without_inventory_tracking
-    checkout_info = fetch_checkout_info(
-        checkout, [], get_plugins_manager(allow_replica=False)
-    )
-    add_variant_to_checkout(checkout_info, variant, 1)
-    checkout.shipping_address = address
-    checkout.shipping_method = shipping_method
-    checkout.billing_address = address
-    checkout.metadata_storage.store_value_in_metadata(items={"accepted": "true"})
-    checkout.metadata_storage.store_value_in_private_metadata(
-        items={"accepted": "false"}
-    )
-    checkout.save()
-    checkout.metadata_storage.save()
-    return checkout
-
-
-@pytest.fixture
-def checkout_with_variants(
-    checkout,
-    stock,
-    product_with_default_variant,
-    product_with_single_variant,
-    product_with_two_variants,
-):
-    checkout_info = fetch_checkout_info(
-        checkout, [], get_plugins_manager(allow_replica=False)
-    )
-
-    add_variant_to_checkout(
-        checkout_info, product_with_default_variant.variants.get(), 1
-    )
-    add_variant_to_checkout(
-        checkout_info, product_with_single_variant.variants.get(), 10
-    )
-    add_variant_to_checkout(
-        checkout_info, product_with_two_variants.variants.first(), 3
-    )
-    add_variant_to_checkout(checkout_info, product_with_two_variants.variants.last(), 5)
-
-    checkout.save()
-    return checkout
-
-
-@pytest.fixture
-def checkout_with_shipping_address(checkout_with_variants, address):
-    checkout = checkout_with_variants
-
-    checkout.shipping_address = address.get_copy()
-    checkout.save()
-
-    return checkout
-
-
-@pytest.fixture
-def checkout_with_variants_for_cc(
-    checkout, stocks_for_cc, product_variant_list, product_with_two_variants
-):
-    CheckoutLine.objects.bulk_create(
-        [
-            CheckoutLine(
-                checkout=checkout,
-                variant=product_variant_list[0],
-                quantity=3,
-                currency="USD",
-            ),
-            CheckoutLine(
-                checkout=checkout,
-                variant=product_variant_list[1],
-                quantity=10,
-                currency="USD",
-            ),
-            CheckoutLine(
-                checkout=checkout,
-                variant=product_with_two_variants.variants.last(),
-                quantity=5,
-                currency="USD",
-            ),
-        ]
-    )
-    return checkout
-
-
-@pytest.fixture
-def checkout_with_shipping_address_for_cc(checkout_with_variants_for_cc, address):
-    checkout = checkout_with_variants_for_cc
-
-    checkout.shipping_address = address.get_copy()
-    checkout.save()
-
-    return checkout
-
-
-@pytest.fixture
-def checkout_with_items(checkout, product_list, product):
-    variant = product.variants.get()
-    checkout_info = fetch_checkout_info(
-        checkout, [], get_plugins_manager(allow_replica=False)
-    )
-    add_variant_to_checkout(checkout_info, variant, 1)
-    for prod in product_list:
-        variant = prod.variants.get()
-        add_variant_to_checkout(checkout_info, variant, 1)
-    checkout.save()
-    checkout.refresh_from_db()
-    return checkout
-
-
-@pytest.fixture
-def checkout_with_items_and_shipping(checkout_with_items, address, shipping_method):
-    checkout_with_items.shipping_address = address
-    checkout_with_items.shipping_method = shipping_method
-    checkout_with_items.billing_address = address
-    checkout_with_items.save()
-    return checkout_with_items
-
-
-@pytest.fixture
-def checkout_with_item_and_shipping(checkout_with_item, address, shipping_method):
-    checkout_with_item.shipping_address = address
-    checkout_with_item.shipping_method = shipping_method
-    checkout_with_item.billing_address = address
-    checkout_with_item.save()
-    return checkout_with_item
-
-
-@pytest.fixture
-def checkout_with_voucher(checkout, product, voucher):
-    variant = product.variants.get()
-    checkout_info = fetch_checkout_info(
-        checkout, [], get_plugins_manager(allow_replica=False)
-    )
-    add_variant_to_checkout(checkout_info, variant, 3)
-    checkout.voucher_code = voucher.code
-    checkout.discount = Money("20.00", "USD")
-    checkout.save()
-    return checkout
-
-
-@pytest.fixture
-def checkout_with_voucher_percentage(checkout, product, voucher_percentage):
-    variant = product.variants.get()
-    checkout_info = fetch_checkout_info(
-        checkout, [], get_plugins_manager(allow_replica=False)
-    )
-    add_variant_to_checkout(checkout_info, variant, 3)
-    checkout.voucher_code = voucher_percentage.code
-    checkout.discount = Money("3.00", "USD")
-    checkout.save()
-    return checkout
-
-
-@pytest.fixture
-def checkout_with_voucher_free_shipping(
-    checkout_with_items_and_shipping, voucher_free_shipping
-):
-    manager = get_plugins_manager(allow_replica=False)
-    lines, _ = fetch_checkout_lines(checkout_with_items_and_shipping)
-    checkout_info = fetch_checkout_info(
-        checkout_with_items_and_shipping, lines, manager
-    )
-    add_voucher_to_checkout(
-        manager,
-        checkout_info,
-        lines,
-        voucher_free_shipping,
-        voucher_free_shipping.codes.first(),
-    )
-    return checkout_with_items_and_shipping
-
-
-@pytest.fixture
-def checkout_with_gift_card(checkout_with_item, gift_card):
-    checkout_with_item.gift_cards.add(gift_card)
-    checkout_with_item.save()
-    return checkout_with_item
-
-
-@pytest.fixture
-def checkout_with_preorders_only(
-    checkout,
-    stocks_for_cc,
-    preorder_variant_with_end_date,
-    preorder_variant_channel_threshold,
-):
-    lines, _ = fetch_checkout_lines(checkout)
-    checkout_info = fetch_checkout_info(
-        checkout, lines, get_plugins_manager(allow_replica=False)
-    )
-    add_variant_to_checkout(checkout_info, preorder_variant_with_end_date, 2)
-    add_variant_to_checkout(checkout_info, preorder_variant_channel_threshold, 2)
-
-    checkout.save()
-    return checkout
-
-
-@pytest.fixture
-def checkout_with_preorders_and_regular_variant(
-    checkout, stocks_for_cc, preorder_variant_with_end_date, product_variant_list
-):
-    lines, _ = fetch_checkout_lines(checkout)
-    checkout_info = fetch_checkout_info(
-        checkout, lines, get_plugins_manager(allow_replica=False)
-    )
-    add_variant_to_checkout(checkout_info, preorder_variant_with_end_date, 2)
-    add_variant_to_checkout(checkout_info, product_variant_list[0], 2)
-
-    checkout.save()
-    return checkout
-
-
-@pytest.fixture
-def checkout_with_gift_card_items(
-    checkout, non_shippable_gift_card_product, shippable_gift_card_product
-):
-    checkout_info = fetch_checkout_info(
-        checkout, [], get_plugins_manager(allow_replica=False)
-    )
-    non_shippable_variant = non_shippable_gift_card_product.variants.get()
-    shippable_variant = shippable_gift_card_product.variants.get()
-    add_variant_to_checkout(checkout_info, non_shippable_variant, 1)
-    add_variant_to_checkout(checkout_info, shippable_variant, 2)
-    checkout.save()
-    return checkout
-
-
-@pytest.fixture
-def checkout_with_voucher_percentage_and_shipping(
-    checkout_with_voucher_percentage, shipping_method, address
-):
-    checkout = checkout_with_voucher_percentage
-    checkout.shipping_method = shipping_method
-    checkout.shipping_address = address
-    checkout.save()
-    return checkout
-
-
-@pytest.fixture
-def checkout_with_payments(checkout):
-    Payment.objects.bulk_create(
-        [
-            Payment(
-                gateway="mirumee.payments.dummy", is_active=True, checkout=checkout
-            ),
-            Payment(
-                gateway="mirumee.payments.dummy", is_active=False, checkout=checkout
-            ),
-        ]
-    )
-    return checkout
-
-
-@pytest.fixture
-def checkout_with_item_and_preorder_item(
-    checkout_with_item, product, preorder_variant_channel_threshold
-):
-    checkout_info = fetch_checkout_info(
-        checkout_with_item, [], get_plugins_manager(allow_replica=False)
-    )
-    add_variant_to_checkout(checkout_info, preorder_variant_channel_threshold, 1)
-    return checkout_with_item
-
-
-@pytest.fixture
-def checkout_with_problems(
-    checkout_with_items,
-    product_type,
-    address,
-    shipping_method,
-    category,
-    default_tax_class,
-    channel_USD,
-    warehouse,
-):
-    checkout_with_items.shipping_address = address
-    checkout_with_items.billing_address = address
-    checkout_with_items.shipping_method = shipping_method
-    checkout_with_items.save(
-        update_fields=["shipping_address", "shipping_method", "billing_address"]
-    )
-
-    first_line = checkout_with_items.lines.first()
-    first_line.variant.track_inventory = True
-    first_line.variant.save(update_fields=["track_inventory"])
-
-    product_type = first_line.variant.product.product_type
-    product_type.is_shipping_required = True
-    product_type.save(update_fields=["is_shipping_required"])
-
-    first_line.variant.stocks.all().delete()
-
-    second_line = checkout_with_items.lines.last()
-
-    available_at = datetime.datetime.now(tz=datetime.UTC) + datetime.timedelta(days=5)
-    product = second_line.variant.product
-    product.channel_listings.update(
-        available_for_purchase_at=available_at, is_published=False
-    )
-
-    return checkout_with_items
 
 
 @pytest.fixture
@@ -996,81 +273,6 @@ def customer_users(address, customer_user, customer_user2):
     customer_user3._password = "password"
 
     return [customer_user, customer_user2, customer_user3]
-
-
-@pytest.fixture
-def user_checkout(customer_user, channel_USD):
-    checkout = Checkout.objects.create(
-        user=customer_user,
-        channel=channel_USD,
-        billing_address=customer_user.default_billing_address,
-        shipping_address=customer_user.default_shipping_address,
-        note="Test notes",
-        currency="USD",
-    )
-    CheckoutMetadata.objects.create(checkout=checkout)
-    return checkout
-
-
-@pytest.fixture
-def user_checkout_for_cc(customer_user, channel_USD, warehouse_for_cc):
-    checkout = Checkout.objects.create(
-        user=customer_user,
-        email=customer_user.email,
-        channel=channel_USD,
-        billing_address=customer_user.default_billing_address,
-        shipping_address=warehouse_for_cc.address,
-        collection_point=warehouse_for_cc,
-        note="Test notes",
-        currency="USD",
-    )
-    return checkout
-
-
-@pytest.fixture
-def user_checkout_PLN(customer_user, channel_PLN):
-    checkout = Checkout.objects.create(
-        user=customer_user,
-        channel=channel_PLN,
-        billing_address=customer_user.default_billing_address,
-        shipping_address=customer_user.default_shipping_address,
-        note="Test notes",
-        currency="PLN",
-    )
-    return checkout
-
-
-@pytest.fixture
-def user_checkout_with_items(user_checkout, product_list):
-    checkout_info = fetch_checkout_info(
-        user_checkout, [], get_plugins_manager(allow_replica=False)
-    )
-    for product in product_list:
-        variant = product.variants.get()
-        add_variant_to_checkout(checkout_info, variant, 1)
-    user_checkout.refresh_from_db()
-    return user_checkout
-
-
-@pytest.fixture
-def user_checkout_with_items_for_cc(user_checkout_for_cc, product_list):
-    checkout_info = fetch_checkout_info(
-        user_checkout_for_cc, [], get_plugins_manager(allow_replica=False)
-    )
-    for product in product_list:
-        variant = product.variants.get()
-        add_variant_to_checkout(checkout_info, variant, 1)
-    user_checkout_for_cc.refresh_from_db()
-    return user_checkout_for_cc
-
-
-@pytest.fixture
-def user_checkouts(request, user_checkout_with_items, user_checkout_with_items_for_cc):
-    if request.param == "regular":
-        return user_checkout_with_items
-    if request.param == "click_and_collect":
-        return user_checkout_with_items_for_cc
-    raise ValueError("Internal test error")
 
 
 @pytest.fixture
@@ -1856,95 +1058,6 @@ def image_list():
         SimpleUploadedFile("image1.jpg", img_data_1.getvalue()),
         SimpleUploadedFile("image2.jpg", img_data_2.getvalue()),
     ]
-
-
-@pytest.fixture
-def checkout_line_with_reservation_in_many_stocks(
-    customer_user, variant_with_many_stocks, checkout
-):
-    variant = variant_with_many_stocks
-    stocks = variant.stocks.all().order_by("pk")
-    checkout_line = checkout.lines.create(
-        variant=variant,
-        quantity=3,
-    )
-
-    reserved_until = timezone.now() + datetime.timedelta(minutes=5)
-
-    Reservation.objects.bulk_create(
-        [
-            Reservation(
-                checkout_line=checkout_line,
-                stock=stocks[0],
-                quantity_reserved=2,
-                reserved_until=reserved_until,
-            ),
-            Reservation(
-                checkout_line=checkout_line,
-                stock=stocks[1],
-                quantity_reserved=1,
-                reserved_until=reserved_until,
-            ),
-        ]
-    )
-
-    return checkout_line
-
-
-@pytest.fixture
-def checkout_line_with_one_reservation(
-    customer_user, variant_with_many_stocks, checkout
-):
-    variant = variant_with_many_stocks
-    stocks = variant.stocks.all().order_by("pk")
-    checkout_line = checkout.lines.create(
-        variant=variant,
-        quantity=2,
-    )
-
-    reserved_until = timezone.now() + datetime.timedelta(minutes=5)
-
-    Reservation.objects.create(
-        checkout_line=checkout_line,
-        stock=stocks[0],
-        quantity_reserved=2,
-        reserved_until=reserved_until,
-    )
-
-    return checkout_line
-
-
-@pytest.fixture
-def checkout_line_with_preorder_item(
-    checkout, product, preorder_variant_channel_threshold
-):
-    checkout_info = fetch_checkout_info(
-        checkout, [], get_plugins_manager(allow_replica=False)
-    )
-    add_variant_to_checkout(checkout_info, preorder_variant_channel_threshold, 1)
-    return checkout.lines.last()
-
-
-@pytest.fixture
-def checkout_line_with_reserved_preorder_item(
-    checkout, product, preorder_variant_channel_threshold
-):
-    checkout_info = fetch_checkout_info(
-        checkout, [], get_plugins_manager(allow_replica=False)
-    )
-    add_variant_to_checkout(checkout_info, preorder_variant_channel_threshold, 2)
-    checkout_line = checkout.lines.last()
-
-    reserved_until = timezone.now() + datetime.timedelta(minutes=5)
-
-    PreorderReservation.objects.create(
-        checkout_line=checkout_line,
-        product_variant_channel_listing=checkout_line.variant.channel_listings.first(),
-        quantity_reserved=2,
-        reserved_until=reserved_until,
-    )
-
-    return checkout_line
 
 
 @pytest.fixture
@@ -2848,140 +1961,6 @@ def delivery_method(request, warehouse_for_cc, shipping_method):
     if request.param == "shipping_method":
         return shipping_method
     return None
-
-
-@pytest.fixture
-def checkout_for_cc(channel_USD, customer_user):
-    checkout = Checkout.objects.create(
-        channel=channel_USD,
-        billing_address=customer_user.default_billing_address,
-        shipping_address=customer_user.default_shipping_address,
-        note="Test notes",
-        currency="USD",
-        email=customer_user.email,
-    )
-    CheckoutMetadata.objects.create(checkout=checkout)
-    return checkout
-
-
-@pytest.fixture
-def checkout_with_items_for_cc(checkout_for_cc, product_variant_list):
-    CheckoutLine.objects.bulk_create(
-        [
-            CheckoutLine(
-                checkout=checkout_for_cc,
-                variant=product_variant_list[0],
-                quantity=1,
-                currency=checkout_for_cc.currency,
-            ),
-            CheckoutLine(
-                checkout=checkout_for_cc,
-                variant=product_variant_list[1],
-                quantity=1,
-                currency=checkout_for_cc.currency,
-            ),
-            CheckoutLine(
-                checkout=checkout_for_cc,
-                variant=product_variant_list[2],
-                quantity=1,
-                currency=checkout_for_cc.currency,
-            ),
-        ]
-    )
-    checkout_for_cc.set_country("US", commit=True)
-
-    return checkout_for_cc
-
-
-@pytest.fixture
-def checkout_with_item_for_cc(checkout_for_cc, product_variant_list):
-    CheckoutLine.objects.create(
-        checkout=checkout_for_cc,
-        variant=product_variant_list[0],
-        quantity=1,
-        currency=checkout_for_cc.currency,
-    )
-    return checkout_for_cc
-
-
-@pytest.fixture
-def checkout_with_prices(
-    checkout_with_items,
-    address,
-    address_other_country,
-    warehouse,
-    customer_user,
-    shipping_method,
-    voucher,
-):
-    # Need to save shipping_method before fetching checkout info.
-    checkout_with_items.shipping_method = shipping_method
-    checkout_with_items.save(update_fields=["shipping_method"])
-
-    manager = get_plugins_manager(allow_replica=False)
-    lines = checkout_with_items.lines.all()
-    lines_info, _ = fetch_checkout_lines(checkout_with_items)
-    checkout_info = fetch_checkout_info(checkout_with_items, lines_info, manager)
-
-    for line, line_info in zip(lines, lines_info):
-        line.total_price_net_amount = base_calculations.calculate_base_line_total_price(
-            line_info
-        ).amount
-        line.total_price_gross_amount = line.total_price_net_amount * Decimal("1.230")
-
-    checkout_with_items.discount_amount = Decimal("5.000")
-    checkout_with_items.discount_name = "Voucher 5 USD"
-    checkout_with_items.user = customer_user
-    checkout_with_items.billing_address = address
-    checkout_with_items.shipping_address = address_other_country
-    checkout_with_items.collection_point = warehouse
-    checkout_with_items.subtotal_net_amount = Decimal("100.000")
-    checkout_with_items.subtotal_gross_amount = Decimal("123.000")
-    checkout_with_items.total_net_amount = Decimal("150.000")
-    checkout_with_items.total_gross_amount = Decimal("178.000")
-    shipping_amount = base_calculations.base_checkout_delivery_price(
-        checkout_info, lines_info
-    ).amount
-    checkout_with_items.shipping_price_net_amount = shipping_amount
-    checkout_with_items.shipping_price_gross_amount = shipping_amount * Decimal("1.08")
-    checkout_with_items.metadata_storage.metadata = {"meta_key": "meta_value"}
-    checkout_with_items.metadata_storage.private_metadata = {
-        "priv_meta_key": "priv_meta_value"
-    }
-
-    checkout_with_items.lines.bulk_update(
-        lines,
-        [
-            "total_price_net_amount",
-            "total_price_gross_amount",
-        ],
-    )
-
-    checkout_with_items.save(
-        update_fields=[
-            "discount_amount",
-            "discount_name",
-            "user",
-            "billing_address",
-            "shipping_address",
-            "collection_point",
-            "subtotal_net_amount",
-            "subtotal_gross_amount",
-            "total_net_amount",
-            "total_gross_amount",
-            "shipping_price_net_amount",
-            "shipping_price_gross_amount",
-        ]
-    )
-    checkout_with_items.metadata_storage.save(
-        update_fields=["metadata", "private_metadata"]
-    )
-
-    user = checkout_with_items.user
-    user.metadata = {"user_public_meta_key": "user_public_meta_value"}
-    user.save(update_fields=["metadata"])
-
-    return checkout_with_items
 
 
 @pytest.fixture
